+++ app/Main.hs
@@ -9,5 +9,5 @@ main = display window white render
     render = renderScene (screenWidth, screenHeight)
     window = InWindow windowName (screenWidth, screenHeight) (screenWidth, screenHeight)
     windowName = "funtracer"
-    screenWidth = 256
-    screenHeight = 256
+    screenWidth = 640
+    screenHeight = 360